Rename package settings to oldsettings
