+++ app/src/main/java/dev/patrickgold/florisboard/oldsettings/ThemeManagerActivity.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package dev.patrickgold.florisboard.settings
+package dev.patrickgold.florisboard.oldsettings
 
 import android.content.Intent
 import android.net.Uri
