+++ app/src/main/java/dev/patrickgold/florisboard/oldsettings/components/TimePickerDialogPreference.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package dev.patrickgold.florisboard.settings.components
+package dev.patrickgold.florisboard.oldsettings.components
 
 import android.app.TimePickerDialog
 import android.content.Context
