+++ app/src/main/java/dev/patrickgold/florisboard/oldsettings/fragments/ThemeFragment.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package dev.patrickgold.florisboard.settings.fragments
+package dev.patrickgold.florisboard.oldsettings.fragments
 
 import android.content.SharedPreferences
 import android.os.Bundle
@@ -24,7 +24,7 @@ import androidx.preference.PreferenceFragmentCompat
 import dev.patrickgold.florisboard.R
 import dev.patrickgold.florisboard.ime.core.Preferences
 import dev.patrickgold.florisboard.ime.theme.ThemeMode
-import dev.patrickgold.florisboard.settings.components.ThemeSelectorPreference
+import dev.patrickgold.florisboard.oldsettings.components.ThemeSelectorPreference
 
 class ThemeFragment : PreferenceFragmentCompat(),
     SharedPreferences.OnSharedPreferenceChangeListener {
