+++ app/src/main/java/dev/patrickgold/florisboard/oldsettings/spelling/DictionarySourcesInfoFragment.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package dev.patrickgold.florisboard.settings.spelling
+package dev.patrickgold.florisboard.oldsettings.spelling
 
 import android.content.Intent
 import android.net.Uri
