+++ app/src/main/res/layout/advanced_activity.xml
@@ -8,7 +8,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/advanced_fragment_frame"
-        android:name="dev.patrickgold.florisboard.settings.fragments.AdvancedFragment"
+        android:name="dev.patrickgold.florisboard.oldsettings.fragments.AdvancedFragment"
         android:layout_width="match_parent"
         android:layout_height="0dp"
         android:layout_weight="1"/>
