+++ app/src/main/res/layout/theme_editor_activity.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
     android:theme="@style/SettingsTheme"
-    tools:context=".settings.ThemeEditorActivity">
+    tools:context=".oldsettings.ThemeEditorActivity">
 
     <include layout="@layout/toolbar"/>
