+++ app/src/main/res/layout/theme_manager_activity.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
     android:theme="@style/SettingsTheme"
-    tools:context=".settings.ThemeManagerActivity">
+    tools:context=".oldsettings.ThemeManagerActivity">
 
     <include layout="@layout/toolbar"/>
